+++ crates/rustc_codegen_spirv/src/abi.rs
@@ -1019,6 +1019,7 @@ fn trans_intrinsic_type<'tcx>(
         IntrinsicType::AccelerationStructureKhr => {
             Ok(SpirvType::AccelerationStructureKhr.def(span, cx))
         }
+
         IntrinsicType::RayQueryKhr => Ok(SpirvType::RayQueryKhr.def(span, cx)),
         IntrinsicType::CooperativeMatrixKhr => {
             if ty.size != Size::from_bytes(4) {