+++ crates/rustc_codegen_spirv/src/builder/mod.rs
@@ -1,5 +1,6 @@
 mod builder_methods;
 mod byte_addressable_buffer;
+mod descriptor_heap;
 mod ext_inst;
 mod format_args_decompiler;
 mod intrinsics;