+++ crates/rustc_codegen_spirv/src/linker/inline.rs
@@ -861,7 +861,7 @@ impl Inliner<'_, '_> {
                                         current_debug_src_loc_inst = callsite_debug_src_loc_inst;
                                     }
                                 }
-                                CustomOp::Abort => break,
+                                CustomOp::Abort | CustomOp::DescriptorHeapLoad => break,
                             }
                         }
                         Op::Variable => {}
@@ -1109,7 +1109,7 @@ impl Inliner<'_, '_> {
                             inlined_frames_depth = inlined_frames_depth.saturating_sub(1);
                             continue;
                         }
-                        CustomOp::Abort => break,
+                        CustomOp::Abort | CustomOp::DescriptorHeapLoad => break,
                     }
                 }
                 Op::Variable => continue,